refactor: move lib/editor-adapter → adapters/editor-adapter (stage 5)
editor-adapter is TypeScript, not a MoonBit module, and its presence in lib/
caused category confusion. Moved to adapters/ with git mv so history is
preserved. All 6 TS consumers (examples/prosemirror/*, examples/web/*) now
import from the new path; pure depth-preserving substitution.

docs/development/ADDING_A_LANGUAGE.md updated to reflect the new path.
Historical plan docs (2026-04-01, 2026-04-18) left unchanged — they describe
the state at that time.

lib/ now contains only MoonBit modules: btree, text-change, semantic, zipper
(+ semantic/proof).
