+++ adapters/editor-adapter/adapter.ts
