+++ adapters/editor-adapter/block-input.css
