+++ adapters/editor-adapter/block-input.ts
