+++ adapters/editor-adapter/cm6-adapter.ts
