+++ adapters/editor-adapter/html-adapter.ts
