+++ adapters/editor-adapter/index.ts
