+++ adapters/editor-adapter/types.ts
