+++ examples/prosemirror/src/keymap.ts
@@ -1,6 +1,6 @@
 import { keymap } from "prosemirror-keymap";
 import { NodeSelection } from "prosemirror-state";
-import type { UserIntent } from "../../../lib/editor-adapter";
+import type { UserIntent } from "../../../adapters/editor-adapter";
 
 /**
  * ProseMirror keymap plugin for structural operations on AST nodes.
